fix: resolve 6 spec-compliance and interop bugs with regression tests
Fix six bugs found during comprehensive SOME/IP spec audit:

1. SD ServiceEntry minor_version truncated to uint8_t (#245)
   - Changed to uint32_t, serialize/deserialize all 4 bytes (12-15)
   - Per SOME/IP-SD spec, Minor Version is a 32-bit field

2. SD ConfigurationOption length off-by-one (#246)
   - Length field now includes Reserved byte: 1 + config_string_.size()
   - Deserialize reads length_ - 1 bytes after reserved consumed

3. SD unknown option skip off-by-one (#247)
   - Changed from offset += 4 + len to offset += 3 + len
   - Total option = Length(2) + Type(1) + length_value

4. E2E default profile_name mismatch (#248)
   - E2EConfig::profile_name default changed "standard" -> "basic"
   - Matches BasicE2EProfile::get_profile_name()

5. Serializer array length prefix semantic mismatch (#249)
   - serialize_array now writes byte length (not element count)
   - Uses back-fill pattern for correct variable-length support

6. uint64 ser/des broken on big-endian targets (#250)
   - Replaced swap-and-memcpy with portable MSB-first extraction
   - Correct on both little-endian and big-endian hosts

Each bug has a corresponding GitHub issue and regression test.
Existing tests updated to match corrected spec-compliant behavior.
All 15 test suites pass (100%).

Closes #245, closes #246, closes #247, closes #248, closes #249, closes #250

+++ include/sd/sd_message.h
@@ -75,8 +75,8 @@ public:
     uint8_t get_major_version() const { return major_version_; }
     void set_major_version(uint8_t version) { major_version_ = version; }
 
-    uint8_t get_minor_version() const { return minor_version_; }
-    void set_minor_version(uint8_t version) { minor_version_ = version; }
+    uint32_t get_minor_version() const { return minor_version_; }
+    void set_minor_version(uint32_t version) { minor_version_ = version; }
 
     std::vector<uint8_t> serialize() const override;
     bool deserialize(const std::vector<uint8_t>& data, size_t& offset) override;
@@ -85,7 +85,7 @@ private:
     uint16_t service_id_{0};
     uint16_t instance_id_{0};
     uint8_t major_version_{0};
-    uint8_t minor_version_{0};
+    uint32_t minor_version_{0};
 };
 
 /**
